+++ src/components/molecules/index.js
@@ -1,1 +1,2 @@
 export {GeneralHeader} from './GeneralHeader';
+export {SignButtons} from './SignButtons';
