+++ src/screens/Register.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import {View, StyleSheet} from 'react-native';
 import {GeneralLink, GeneralText} from '../components/atoms';
-import {GeneralHeader} from '../components/molecules/index';
+import {GeneralHeader, SignButtons} from '../components/molecules/index';
 import inputsMapping from '../utils/inputsMapping';
-import buttonsMapping from '../utils/buttonsMapping';
 
 const inputsRender = () => {
   const generalInputs = [
@@ -28,37 +27,20 @@ const inputsRender = () => {
   return inputsMapping(generalInputs);
 };
 
-const buttonsRender = navigation => {
-  const generalButtons = [
-    {
-      title: 'Registrarse',
-      color: 'primary',
-      action: () => navigation.navigate('Path'),
-    },
-    {
-      title: 'Registrarse con Facebook',
-      color: 'facebook',
-      icon: 'logo-facebook',
-    },
-  ];
-  return buttonsMapping(generalButtons);
-};
-
 export const Register = ({navigation}) => {
   return (
     <>
       <GeneralHeader />
       <View style={styles.registro}>
-        <GeneralText
-          size={'h1'}
-          color={'secondary'}
-          title={'Regístrate'}
-          weight
-        />
+        <GeneralText size={'h1'} color={'secondary'} title={'Regístrate'} />
       </View>
       <View style={styles.center}>
         {inputsRender()}
-        <View style={styles.botones}>{buttonsRender(navigation)}</View>
+        <SignButtons
+          style={styles.botones}
+          navigation={navigation}
+          type="register"
+        />
         <GeneralLink
           title={'¿Ya eres usuario? Inicia sesión'}
           size={'h4'}
